requirements.txt
pandas==2.2.2
psycopg2-binary==2.9.9
sqlalchemy==2.0.30
python-dotenv==1.0.1
tqdm==4.66.4
